# Coursera-NIIT/0x01-building-restful-apis-using-nodejs-and-express/week-2/movie-application-master/api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: Movie API
  version: 1.0.0
  contact:
    email: phyrokelstein@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: "https://8000-2022phyro-courseraniit-wql4h556u3v.ws-eu107.gitpod.io/api/v1"
  - url: "http:localhost:3000/api/v1"
paths:
  /movies:
    get:
      tags: [Movie details]
      summary: Get details of all movies
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                STATUS: OK
                data:
                  - id: 2
                    movieName: The Guard
                    director: John Michael McDonagh
                    rating: 9.4

    post:
      tags: [Movie details]
      summary: Save new movie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                STATUS: OK
                data:
                  id: 2
                  movieName: The Guard
                  director: John Michael McDonagh
                  rating: 9.4

  /movies/{movieId}:
    parameters:
      - name: movieId
        in: path
        required: true
        schema:
          type: integer
          example: 2
    get:
      tags: [Movie details]
      summary: Get details of a specific movie
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                STATUS: OK
                data:
                  id: 2
                  movieName: The Guard
                  director: John Michael McDonagh
                  rating: 9.4
        '404':
          description: Movie not found

    patch:
      tags: [Movie details]
      summary: Edit movie detail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                STATUS: OK
                data:
                  id: 2
                  movieName: Updated Movie
                  director: John Michael McDonagh
                  rating: 9.4
        '404':
          description: Movie not found

    delete:
      tags: [Movie details]
      summary: Delete movie
      responses:
        '204':
          description: No Content
        '404':
          description: Movie not found
tags:
  - name: Movie details
    description: The documentation for the movies
components:
  schemas:
    Movie:
      type: object
      properties:
        id:
          type: integer
          example: 2
        movieName:
          type: string
          example: Shawshanks Redemption
        director:
          type: string
          example: Steven Spielberg
        rating:
          type: number
          format: float
          example: 9.4